% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_Deconv.R
\docType{data}
\name{Box_Deconv}
\alias{Box_Deconv}
\alias{my_comparisons}
\title{Box_Deconv
Box_Deconv

Function that generates a boxplot comparing two conditions for each cell type from a df and a condition vector.}
\format{
An object of class \code{list} of length 3.
}
\usage{
my_comparisons
}
\arguments{
\item{data}{Data Frame resulting from the deconvolution.}

\item{cond}{Vector that assigns a condition to each sample of the data frame.}

\item{results_dir}{Directory to which the graphics will be generated.}

\item{f_name}{Name you want to give to the file, by default it is "Deconvolution Boxplot".}
}
\value{
The function returns the boxplots graph in the specified directory.
}
\description{
Box_Deconv
Box_Deconv

Function that generates a boxplot comparing two conditions for each cell type from a df and a condition vector.
}
\examples{
Box_Deconv(data= C_abs, cond= fractions, results_dir = "/bicoh/nidia/Deconv/Brain_MMartin/test", f_name= "Box_CIBERSORT_abs")
}
\author{
Nidia Barco Armengol
}
\keyword{datasets}
