% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT_rel_F.R
\name{CIBERSORT_rel_F}
\alias{CIBERSORT_rel_F}
\title{CIBERSORT_rel_F}
\usage{
CIBERSORT_rel_F(
  matrix,
  sig.matrix,
  fractions,
  results_dir,
  height = 10,
  name = FALSE
)
}
\arguments{
\item{matrix}{expression matrix, this matrix must be in .txt format, organized in tabulated columns with one column containing the list of genes, with the name Gene.symbol, Gene.Symbol or Geneid, and in the rest of the columns the names of the samples and the expression values for the genes.  (Example. Counts.HUMAN_nodup_Sal_Veh_IPs_INPUTs.txt).}

\item{sig.matrix}{Signature matrix, this matrix must be in .txt format, organized in tabulated columns with one column containing the list of genes, with the name Gene.symbol, Gene.Symbol or Geneid, and in the rest of the columns the cell types and the expression values for the genes.  (Example. Monaco-ABIS_Immune11_Microarray.txt).}

\item{fractions}{Vector that identifies the fractions of the sample list to be used to fractionate the samples in different graphs.}

\item{results_dir}{Directory to which the graphics will be generated.}

\item{height}{Number that defines the height of the graph, to increase or decrease the height, which has a default value of 10.}

\item{name}{Alternative name for the chart, which by default has the structure Sig_type_method.matrixName_Plot.png (Example. "graph_1")}
}
\value{
Returns the generated graphs and the df of the deconvolution.
}
\description{
Function that generates the relative deconvolution plot with the CIBERSORT method, with the fractional plot, and returns a df resulting from the deconvolution.
}
\examples{
CIBERSORT_rel_F("Counts.HUMAN_nodup_Sal_Veh_IPs_INPUTs.txt", "Monaco-ABIS_Immune11_Microarray.txt", f, results_dir = "/bicoh/nidia/Deconv/Brain_MMartin/test", height= 8, name= "MONACO")
}
\author{
Nidia Barco Armengol
}
