% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{cspline_utils}
\alias{cspline_utils}
\title{generate utility matrices for fitting penalized cubic spline regressions}
\usage{
cspline_utils(knots)
}
\arguments{
\item{knots}{unique knots for the cubic spline}
}
\value{
a list with knots (x), gap (h), Q matrix, R matrix and R^-1Q^t
}
\description{
generate utility matrices for fitting penalized cubic spline regressions
}
