% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution_bayes.R
\name{deconv_bayes_2}
\alias{deconv_bayes_2}
\title{Bayesian Inference with two possible precisions for all genes}
\usage{
deconv_bayes_2(
  X,
  Y,
  tau_eta = 0.1,
  alpha = 1,
  beta = 5,
  kappa1 = 10,
  lambda1 = 1,
  kappa2 = 1,
  lambda2 = 3,
  n.chains = 5,
  n.iter = 50000,
  n.burnin = 20000,
  n.thin = 4
)
}
\arguments{
\item{X}{training data(nsamples*ngenes)}

\item{Y}{test data (vector with ngenes)}

\item{tau_eta}{precision for normal distribution of etas (logit link of probabilities)}

\item{alpha}{beta distribution parameter for the proportion of marker genes}

\item{beta}{beta distribution parameter for the proportion of marker genes}

\item{kappa1}{gamma distribution parameter for the precision of useful genes}

\item{lambda1}{gamma distribution parameter for the precision of useful genes}

\item{kappa2}{gamma distribution parameter for the precision of nuisance genes}

\item{lambda2}{gamma distribution parameter for the precision of nuisance genes}

\item{n.chains}{number of chains for MCMC}

\item{n.iter}{number of iterations in each chain}

\item{n.burnin}{number of burnins}

\item{n.thin}{thinning for posterior sampling}
}
\value{
posterior mean of each parameters and mixing check
}
\description{
Bayesian Inference with two possible precisions for all genes
}
