% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution based on l1 loss}
\usage{
deconvolution(
  X,
  Y,
  labels,
  penalty = 0,
  log = T,
  type = c("l1", "l2"),
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{penalty}{penalty coefficient to force weights of similar samples to be similar}

\item{log}{whether to take log of the values}

\item{type}{l1 loss or l2 loss}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l1 loss
}
