% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline_utils.R
\name{deconvolution_fixed_peak}
\alias{deconvolution_fixed_peak}
\title{Deconvolution where the position with maximum weight is predetermined}
\usage{
deconvolution_fixed_peak(
  X,
  Y,
  weights = NULL,
  labels,
  log = T,
  min_scale = 0.01,
  peak,
  verbose = FALSE
)
}
\arguments{
\item{X}{training data matrix (nsample * nfeature)}

\item{Y}{test data vector (nfeature)}

\item{weights}{vector of weights for each feature, if null then is equally one}

\item{labels}{labels for all the training samples, assumed to be unique}

\item{log}{whether to take log before fitting deconvolution}

\item{min_scale}{features whose IQR value smaller than min_scale will be removed}

\item{peak}{the label value with maximum weight}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
weight bases
}
\description{
Deconvolution where the position with maximum weight is predetermined
}
