% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{dmat_utils}
\alias{dmat_utils}
\title{Design matrix for regression given the knots}
\usage{
dmat_utils(t, x)
}
\arguments{
\item{t}{input independent variable values vector}

\item{x}{vector of knots}
}
\description{
Design matrix for regression given the knots
}
