% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{dmat_utils_d2}
\alias{dmat_utils_d2}
\title{Design matrix for regression given the knots (predicting 1st derivative)}
\usage{
dmat_utils_d2(t, x)
}
\arguments{
\item{t}{input independent variable values vector}

\item{x}{vector of knots}
}
\value{
a design matrix with number of rows equal to length of t and number of columns equal to (2*nknots-2)
}
\description{
Design matrix for regression given the knots (predicting 1st derivative)
}
