% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prediction.R
\name{eigen_posterior}
\alias{eigen_posterior}
\title{Given posterior densities fo all the valid features, generate a consensus posterior density and use it for final prediction}
\usage{
eigen_posterior(
  post_density,
  feature_significance = rep(1, nrow(post_density)),
  validity,
  labels
)
}
\arguments{
\item{post_density}{posterior density matrix, nfeature * nlabels, each row sum to one}

\item{feature_significance}{a vector with length nfeature representing the usefulness of each feature, based on F statistic of cubic spline fit}

\item{validity}{a binary vector with length equal to number of features}

\item{labels}{label values}
}
\value{
posterior probability, MAP, posterior mean and the importance of each feature
}
\description{
Given posterior densities fo all the valid features, generate a consensus posterior density and use it for final prediction
}
