% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticnet.R
\name{elasticnet}
\alias{elasticnet}
\title{Elastic Net Regression Prediction}
\usage{
elasticnet(X, Y, labels, log = T, verbose = FALSE)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{log}{whether to take log of the values}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated coefficients and prediction
}
\description{
Elastic Net Regression Prediction
}
