% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_rel_NF.R
\name{EPIC_rel_NF}
\alias{EPIC_rel_NF}
\title{EPIC_rel_NF}
\usage{
EPIC_rel_NF(matrix, sig.matrix, results_dir, height = 10, name = FALSE)
}
\arguments{
\item{matrix}{expression matrix, this matrix must be in .txt format, organized in tabulated columns with one column containing the list of genes, with the name Gene.symbol, Gene.Symbol or Geneid, and in the rest of the columns the names of the samples and the expression values for the genes.  (Example. Counts.HUMAN_nodup_Sal_Veh_IPs_INPUTs.txt).}

\item{sig.matrix}{Signature matrix, this matrix must be in .txt format, organized in tabulated columns with one column containing the list of genes, with the name Gene.symbol, Gene.Symbol or Geneid, and in the rest of the columns the cell types and the expression values for the genes.  (Example. Monaco-ABIS_Immune11_Microarray.txt).}

\item{results_dir}{Directory to which the graphics will be generated.}

\item{height}{Number that defines the height of the graph, to increase or decrease the height, which has a default value of 10.}

\item{name}{Alternative name for the chart, which by default has the structure Sig_type_method.matrixName_Plot.png (Example. "graph_1")}
}
\value{
Returns the generated graph and the df of the deconvolution.
}
\description{
Function that generates the relative deconvolution plot with the EPIC method, with the non-fractional plot, and returns a df resulting from the deconvolution.
}
\examples{
EPIC_rel_NF("Counts.HUMAN_nodup_Sal_Veh_IPs_INPUTs.txt", "Monaco-ABIS_Immune11_Microarray.txt", results_dir = "/bicoh/nidia/Deconv/Brain_MMartin/test", height= 8, name= "MONACO")
}
\author{
Nidia Barco Armengol
}
