% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{f_test_filter}
\alias{f_test_filter}
\title{Filter genes that varies a lot between samples with different labels based on ANOVA}
\usage{
f_test_filter(geneexp_df, labels, min_label, max_label)
}
\arguments{
\item{geneexp_df}{gene expression dataframe with samples on rows and features on columns}

\item{labels}{labels of all the samples}

\item{min_label}{minimum label value to be considered for a sample to be included}

\item{max_label}{maximum label value to be considered for a sample to be included}
}
\value{
a list with the selected potential marker genes and the corresponding ANOVA results
}
\description{
Filter genes that varies a lot between samples with different labels based on ANOVA
}
