% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{fit_cspline}
\alias{fit_cspline}
\title{Fit cubic spline given input training dataframe and knots}
\usage{
fit_cspline(t, y, x, alphas = c(2^seq(-5, 4, 1)))
}
\arguments{
\item{t}{independent variable values}

\item{y}{observed values}

\item{x}{knots}

\item{alphas}{penalty parameters to choose from}
}
\description{
Fit cubic spline given input training dataframe and knots
}
