% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{fit_data}
\alias{fit_data}
\title{Fit cubic spline to all the features}
\usage{
fit_data(input_pheno, input_marker_value, output_pheno)
}
\arguments{
\item{input_pheno}{continuous phenotype values of training data}

\item{input_marker_value}{marker values with dimension nfeature * nsample}

\item{output_pheno}{selected phenotype values for prediction of mean and standard error}
}
\value{
list of spline fitting result and predicted mean/standard error
}
\description{
Fit cubic spline to all the features
}
