% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gam_fit.R
\name{gam_fit}
\alias{gam_fit}
\title{Fit penalized cubic regression splines}
\usage{
gam_fit(df, fname, k = 10)
}
\arguments{
\item{df}{input dataframe with two columns "Age" and "Feature"}

\item{fname}{feature name to be shown in the plot title}

\item{k}{number of spline bases, default 10}
}
\value{
effective degrees of freedom, F statistics, p value, visualization
}
\description{
Fit penalized cubic regression splines
}
