% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{group_counts}
\alias{group_counts}
\title{Count the number of samples in each group}
\usage{
group_counts(expression_df, group)
}
\arguments{
\item{expression_df}{gene expression data frame}

\item{group}{the variable name to group the samples}
}
\value{
a vector representing the counts in each group
}
\description{
Count the number of samples in each group
}
