% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{hcluster_genes}
\alias{hcluster_genes}
\title{Hierarchical clustering of genes based on standardized log gene expressions across different labels}
\usage{
hcluster_genes(geneexp_df, labels, subset_genes, min_label, max_label)
}
\arguments{
\item{geneexp_df}{gene expression dataframe with samples on rows and features on columns}

\item{labels}{labels of all samples}

\item{subset_genes}{a subset of genes to carry out hierarchical clustering}

\item{min_label}{only include samples whose labels are larger than or equal to min_label}

\item{max_label}{only include samples whose labels are smaller or equal to max_label}
}
\value{
dendrogram of hierarchical clustering
}
\description{
Hierarchical clustering of genes based on standardized log gene expressions across different labels
}
