% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline_utils.R
\name{ispline_basis_func}
\alias{ispline_basis_func}
\title{Ispline value for a four-knot equally spaced segment}
\usage{
ispline_basis_func(
  x,
  theta,
  delta,
  multiplicity = c(0, 1, 2),
  loc = c("left", "right"),
  normalize = FALSE
)
}
\arguments{
\item{x}{input x vector value}

\item{theta}{value of the smallest knot}

\item{delta}{value of the gap}

\item{multiplicity}{maximium multiplicity at either end of the knots}

\item{loc}{left or right end knot to have more than zero multiplicity}

\item{normalize}{binary indicator, whether to normalize so that maximum value is one}
}
\value{
return a list that contains all the inputs and the corresponding y values
}
\description{
Ispline value for a four-knot equally spaced segment
}
