% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline_utils.R
\name{ispline_basis_setup}
\alias{ispline_basis_setup}
\title{generate I-spline basis components}
\usage{
ispline_basis_setup(minval, maxval, xvals, nknots = 11)
}
\arguments{
\item{minval}{smallest value of the knots}

\item{maxval}{largest value of the knots}

\item{xvals}{x values to be evalutated for spline values}

\item{nknots}{number of knots including the two end points}
}
\value{
a matrix with nknots+1 basis values (each basis has length equal to xvals)
}
\description{
generate I-spline basis components
}
