% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{iter_deconv}
\alias{iter_deconv}
\title{Iterative deconvolution based on l1 or l2 loss}
\usage{
iter_deconv(
  X,
  Y,
  labels,
  lambda1 = 0,
  lambda2 = 0,
  log = T,
  nu = 0.001,
  epsilon = 0.01,
  max_iter = 100,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{lambda1}{penalty parameter for smoothness}

\item{lambda2}{penalty parameter for LASSO shrinkage}

\item{log}{whether to take log of the values}

\item{nu}{stablizing parameter for reweighing features for deconvolution}

\item{epsilon}{tolerance value for stoppage of iterations}

\item{max_iter}{maximum number of iterations}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
Iterative deconvolution based on l1 or l2 loss
}
