% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{l2_solve}
\alias{l2_solve}
\title{deconvolution based on l2 loss}
\usage{
l2_solve(X, Y, labels, p0 = 0, p1 = 0, verbose = FALSE)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{p0}{penalty parameter for similarity of weights for samples with the same label}

\item{p1}{penalty parameter for smoothness of average weights among three adjacent different labels}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l2 loss
}
