% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{likelihood}
\alias{likelihood}
\title{Generate posterior density for each feature given the cubic spline fit result (mean and se)}
\usage{
likelihood(mean_mat, se_mat, y, exclude_factor = 1)
}
\arguments{
\item{mean_mat}{posterior mean matrix, nfeature*ntimestamps}

\item{se_mat}{posterior mean matrix, nfeature*ntimestamps}

\item{y}{feature values of a new sample, a vector of length nfeature}

\item{exclude_factor}{if a feature's value is larger than max_mean+factor\emph{max_se or smaller than
min_mean-factor}max_se, then the feature is considered abnormal and not considered for prediction
@return posterior density for each feature and a validity mask}
}
\description{
Generate posterior density for each feature given the cubic spline fit result (mean and se)
}
