% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{linear_simulator}
\alias{linear_simulator}
\title{Simulate from a linear function}
\usage{
linear_simulator(knots, slope = 0, intercept = 0, noise = 0.1)
}
\arguments{
\item{knots}{x values}

\item{slope}{slope}

\item{intercept}{intercept}

\item{noise}{standard deviation of noise which is added to the signal}
}
\description{
Simulate from a linear function
}
