% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{meanexp}
\alias{meanexp}
\title{Calculate the mean values (maybe grouped by a column representing groups)}
\usage{
meanexp(expression_df, group = NULL)
}
\arguments{
\item{expression_df}{gene expression data frame}

\item{group}{the variable name to group the samples}
}
\value{
a vector of not grouped by any variable and a matrix if grouped by a variable
}
\description{
Calculate the mean values (maybe grouped by a column representing groups)
}
