% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{mle}
\alias{mle}
\title{Given posterior densities fo all the valid features, generate a consensus posterior density and use it for final prediction}
\usage{
mle(lik, feature_significance = rep(1, nrow(lik)), mask, labels)
}
\arguments{
\item{lik}{likelihood values across all labels for each feature}

\item{feature_significance}{a vector with length nfeature representing the usefulness of each feature, based on adjustedR2 of cubic spline fit}

\item{mask}{a binary vector with length equal to number of features, exclude features whose values are beyond the training data range}

\item{labels}{label values}
}
\value{
joint likelihood curve and mle estimate
}
\description{
Given posterior densities fo all the valid features, generate a consensus posterior density and use it for final prediction
}
