% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ispline_utils.R
\name{monotone_bases}
\alias{monotone_bases}
\title{Set up bases to set up probability where there is only one peak and monotonically decrease otherwise}
\usage{
monotone_bases(labels, peak)
}
\arguments{
\item{labels}{labels where weights are calculated}

\item{peak}{location where weights become the largest}
}
\value{
a list with the monotone bases
}
\description{
Set up bases to set up probability where there is only one peak and monotonically decrease otherwise
}
