% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS.R
\name{monotone_check}
\alias{monotone_check}
\title{check monotonicity}
\usage{
monotone_check(ages, expression_df, takelog = T)
}
\arguments{
\item{ages}{ages of all samples}

\item{expression_df}{gene expression data frame (sampe * gene)}

\item{takelog}{whether to take log, by default TRUE}
}
\value{
Regression coefficient and p values
}
\description{
check monotonicity
}
