% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{penalty_smooth}
\alias{penalty_smooth}
\title{Generate penalty matrix}
\usage{
penalty_smooth(labels, lambda = 1)
}
\arguments{
\item{labels}{phenotypes of all the samples}

\item{lambda}{penalty parameter for similarity of weights for samples with the same label}
}
\value{
penalty quadratic matrix
}
\description{
Generate penalty matrix
}
