% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_prediction.R
\name{posterior}
\alias{posterior}
\title{Generate posterior density for each feature given the cubic spline fit result (mean and se)}
\usage{
posterior(
  postmean,
  postse,
  y,
  prior = rep(1, ncol(postse)),
  exclude_factor = 1
)
}
\arguments{
\item{postmean}{posterior mean matrix, nfeature*ntimestamps}

\item{postse}{posterior mean matrix, nfeature*ntimestamps}

\item{y}{feature values of a new sample, a vector of length nfeature}

\item{prior}{prior distribution}

\item{exclude_factor}{if a feature's value is larger than max_mean+factor\emph{max_se or smaller than
min_mean-factor}max_se, then the feature is considered abnormal and not considered for prediction
@return posterior density for each feature and a validity mask}
}
\description{
Generate posterior density for each feature given the cubic spline fit result (mean and se)
}
