% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubic_spline.R
\name{predict_cspline}
\alias{predict_cspline}
\title{predict mean values and standard errors for given input independent variable values}
\usage{
predict_cspline(t, fitted_result)
}
\arguments{
\item{t}{vector of input values}

\item{fitted_result}{a list that contains x (knots of spline), g_hat (values at each knot), sigma_hat2 (estimated random error variance),
proj_mat (used for calculating prediction error)}
}
\description{
predict mean values and standard errors for given input independent variable values
}
