% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{robust_scale}
\alias{robust_scale}
\title{Calculate median value and IQR of each feature}
\usage{
robust_scale(value_mat, margin = 1)
}
\arguments{
\item{value_mat}{data matrix (nsample * nfeature)}

\item{margin}{calculate scale for each row (1) or each column (2)}
}
\value{
median value vector and scale vector
}
\description{
Calculate median value and IQR of each feature
}
