% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{sinusoid_generator}
\alias{sinusoid_generator}
\title{simulate from trignometric functions}
\usage{
sinusoid_generator(
  knots,
  period,
  amplitude = 1,
  phase = 0,
  shift = 0,
  noise = 0.1,
  type = c("sine", "cosine")
)
}
\arguments{
\item{knots}{x values}

\item{period}{period of the trignometric function}

\item{amplitude}{amplitude of thesinusoid function}

\item{phase}{phase of the trignometric function}

\item{shift}{intercept}

\item{noise}{standard deviation of noise which is added to the signal}

\item{type}{sine function or cosine function}
}
\description{
simulate from trignometric functions
}
