% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.R
\name{sum_squares}
\alias{sum_squares}
\title{Calculate sum of squares of error for each column}
\usage{
sum_squares(mat, vec, weights = NULL)
}
\arguments{
\item{mat}{input matrix}

\item{vec}{mean vector}

\item{weights}{weights for different rows}
}
\value{
a vector showing the sum of squares for each column
}
\description{
Calculate sum of squares of error for each column
}
