% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TWAS.R
\name{twas}
\alias{twas}
\title{linear association test}
\usage{
twas(ages, expression_df, takelog = T)
}
\arguments{
\item{ages}{ages of all samples}

\item{expression_df}{gene expression data frame (sampe * gene)}

\item{takelog}{whether to take log, by default TRUE}
}
\value{
Regression coefficient and p values
}
\description{
linear association test
}
