% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{var_comp}
\alias{var_comp}
\title{ANOVA of individual gene expressions}
\usage{
var_comp(expression_df, groups)
}
\arguments{
\item{expression_df}{gene expression data frame}

\item{groups}{a vector of integers indicating group}
}
\value{
a dataframe reflecting the total sum of squares, residual sum of squares, explained sum of squares and F statistic
}
\description{
ANOVA of individual gene expressions
}
