% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_gene}
\alias{viz_gene}
\title{Draw gene expression over age}
\usage{
viz_gene(ages, values, title = "Gene Plot")
}
\arguments{
\item{ages}{labels of training data}

\item{values}{estimated weights}

\item{title}{title of the ggplot}
}
\description{
Draw gene expression over age
}
