% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_predict}
\alias{viz_predict}
\title{Plot the predicted age against the true age}
\usage{
viz_predict(truth, predicted, diagonal = TRUE, title = NULL)
}
\arguments{
\item{truth}{truth vector}

\item{predicted}{vector of predicted values}

\item{diagonal}{whether to add an identity line, default TRUE}

\item{title}{plot title, optional}
}
\description{
Plot the predicted age against the true age
}
