% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{viz_weights}
\alias{viz_weights}
\title{Visualize the deconvolution weights}
\usage{
viz_weights(labels, weights, truth, exclude = NULL, title = NULL)
}
\arguments{
\item{labels}{labels of training data}

\item{weights}{estimated weights}

\item{truth}{true value}

\item{exclude}{which training sample to remove, optional}

\item{title}{title of the ggplot}
}
\description{
Visualize the deconvolution weights
}
