% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution based on l1 or l2 loss}
\usage{
deconvolution(
  X,
  Y,
  labels,
  weights = NULL,
  p0 = 0,
  p1 = 0,
  log = T,
  type = c("l1", "l2"),
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{weights}{vector of weights for each feature, if null then is equally one}

\item{p0}{penalty parameter for similarity of weights for samples with the same label}

\item{p1}{penalty parameter for smoothness of average weights among three adjacent different labels}

\item{log}{whether to take log of the values}

\item{type}{l1 loss or l2 loss}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l1 or l2 loss
}
