% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolution}
\alias{deconvolution}
\title{deconvolution based on l1 or l2 loss}
\usage{
deconvolution(
  X,
  Y,
  labels,
  weights = NULL,
  lambda1 = 0,
  lambda2 = 0,
  alpha = 0.5,
  log = T,
  min_scale = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{weights}{vector of weights for each feature, if null then is equally one}

\item{lambda1}{penalty parameter for smoothness}

\item{lambda2}{penalty parameter for LASSO shrinkage}

\item{alpha}{elastic net parameter to split between ridge and LASSO penalty}

\item{log}{whether to take log of the values}

\item{min_scale}{features whose IQR value smaller than min_scale will be removed}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l1 or l2 loss
}
