% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticnet.R
\name{elasticnet}
\alias{elasticnet}
\title{Elastic Net Regression Prediction}
\usage{
elasticnet(train_data, train_labels, test_data, test_labels)
}
\arguments{
\item{train_data}{feature values of training samples (nsample*nfeature)}

\item{train_labels}{response values of training samples}

\item{test_data}{feature values of test samples (nsample*nfeature)}

\item{test_labels}{response values of test samples}
}
\value{
Estimated coefficients and prediction
}
\description{
Elastic Net Regression Prediction
}
