% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{l2_solve}
\alias{l2_solve}
\title{deconvolution based on l2 loss}
\usage{
l2_solve(
  X,
  Y,
  labels,
  weights = NULL,
  lambda1 = 0,
  lambda2 = 0,
  verbose = FALSE
)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{weights}{vector of weights for each feature, if null then is equally one}

\item{lambda1}{penalty parameter for smoothness}

\item{lambda2}{penalty parameter for LASSO shrinkage}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l2 loss
}
