% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{l2_solve}
\alias{l2_solve}
\title{deconvolution based on l2 loss}
\usage{
l2_solve(X, Y, labels, penalty = 0, verbose = FALSE)
}
\arguments{
\item{X}{matrix of training data (nsample * nfeature)}

\item{Y}{vector of test data}

\item{labels}{labels of all the training samples}

\item{penalty}{penalty coefficient to force weights of similar samples to be similar}

\item{verbose}{Whether to print out details of osqp function, default False}
}
\value{
Estimated weights and the weighted sum of training labels
}
\description{
deconvolution based on l2 loss
}
