% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{penalty_smooth}
\alias{penalty_smooth}
\title{Generate penalty matrix}
\usage{
penalty_smooth(labels, p0 = 1, p1 = 10)
}
\arguments{
\item{labels}{phenotypes of all the samples}

\item{p0}{penalty parameter for similarity of weights for samples with the same label}

\item{p1}{penalty parameter for smoothness of average weights among three adjacent different labels}
}
\value{
penalty quadratic matrix
}
\description{
Generate penalty matrix
}
