% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robust_scale}
\alias{robust_scale}
\title{Calculate median value and IQR of each feature}
\usage{
robust_scale(value_mat)
}
\arguments{
\item{value_mat}{data matrix (nsample * nfeature)}
}
\value{
median value vector and scale vector
}
\description{
Calculate median value and IQR of each feature
}
