% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{scale_transform}
\alias{scale_transform}
\title{Robust scaling of a matrix or a vector}
\usage{
scale_transform(value_mat, median_vals, scale_vals)
}
\arguments{
\item{value_mat}{data matrix (nsample*nfeature) or vector (nfeature)}

\item{median_vals}{median value vector (nfeature)}

\item{scale_vals}{scale value vector (nfeature)}
}
\description{
Robust scaling of a matrix or a vector
}
